requirements.txt
 pydub 
 simpleaudio
sounddevice 
numpy 
scipy
matplotlib 
pillow